+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ TBD
 
 - Fixed SystemClient to revert to actual System Version when using `latest` when validation error occurs when no change in calculated latest. 
   Allowing support for Beer Garden >= 3.26
+- Added support for SystemClient commands to override command type with `_command_type`
 
 3.26.0
 ------